fresh look, folding categories

+++ src/App.tsx
@@ -1,61 +1,89 @@
-import "./App.css";
+import style from "./App.module.css";
 import Packing from "./Packing";
 import Add from "./Add";
-import Edit from "./Edit";
 import { useState } from "react";
 import DataBase from "./db/db";
-import imgEdit from "./assets/edit.png";
+import imgEdit from "./assets/edit.svg";
+import imgAdd from "./assets/add.svg";
+import imgReset from "./assets/reset.svg";
+import imgReturn from "./assets/return.svg";
+import Reset from "./Reset";
 
 function App() {
+  const [stopScroll, setStopScroll] = useState(false);
   const [editMode, setEditMode] = useState(false);
   const [showAdd, setShowAdd] = useState(false);
+  const [showReset, setShowReset] = useState(false);
   const [refresh, setRefresh] = useState(1);
 
+  const toggleEdit = () => {
+    setEditMode((prev) => !prev);
+  };
+
   const onCloseAdd = () => {
+    setStopScroll(false);
     setShowAdd(false);
     setRefresh(refresh + 1);
   };
 
   const onShowAdd = () => {
+    setStopScroll(true);
     setShowAdd(true);
   };
 
-  const onReset = () => {
-    DataBase.delete();
-    location.reload();
+  const onShowReset = () => {
+    setStopScroll(true);
+    setShowReset(true);
+  };
+
+  const onCloseReset = (reset: boolean) => {
+    setStopScroll(false);
+    setShowReset(false);
+    if (reset) {
+      DataBase.delete();
+      location.reload();
+    }
   };
 
   return (
     <>
       {showAdd && (
-        <div className="layer1">
+        <div className={style.layer1}>
           <Add onClose={onCloseAdd}></Add>
         </div>
       )}
-      <div className="layer0">
-        <div className="menu-container">
+      {showReset && (
+        <div className={style.layer1}>
+          <Reset onClose={onCloseReset}></Reset>
+        </div>
+      )}
+      <div className={style.layer0}>
+        <div className={style.menuContainer}>
+          {!editMode && (
+            <img className={style.menu} src={imgEdit} onClick={toggleEdit} />
+          )}
           {editMode && (
             <>
-              <div className="menu reset" onClick={onReset}>
-                Przywróć stan początkowy
-              </div>
-              <div className="menu add" onClick={onShowAdd}>
-                Dodaj nową rzecz
-              </div>
+              <img
+                className={style.menu}
+                src={imgReturn}
+                onClick={toggleEdit}
+              />
+              {<img src={imgAdd} className={style.menu} onClick={onShowAdd} />}
+              {
+                <img
+                  src={imgReset}
+                  className={[style.menu, style.reset].join(" ")}
+                  onClick={onShowReset}
+                />
+              }
             </>
           )}
-          <div
-            className="menu"
-            onClick={() => {
-              setEditMode(!editMode);
-            }}
-          >
-            {!editMode && <img src={imgEdit} width="48px" />}
-            {editMode && "Wróć do pakowania"}
-          </div>
         </div>
-        {editMode && <Edit></Edit>}
-        {!editMode && <Packing></Packing>}
+        <div className={stopScroll ? style.stopScrolling : ""}>
+          {!editMode && <Packing editMode={false}></Packing>}
+          {editMode && <Packing editMode={true}></Packing>}
+        </div>
       </div>
     </>
   );
